Code clearing and refactoring
