+++ app/src/main/res/layout/activity_main.xml
@@ -28,6 +28,6 @@
         android:layout_height="wrap_content"
         android:layout_gravity="bottom|end"
         android:layout_margin="@dimen/fab_margin"
-        app:srcCompat="@android:drawable/ic_dialog_email" />
+        android:src="@drawable/ic_add_black_24dp"/>
 
 </android.support.design.widget.CoordinatorLayout>
